Tweak sliders, modularize, require Java 11
- tweaks sliders with labels
- modularize application (rename package, add module-info)
- technically only requires Java 9
