+++ pom.xml
@@ -13,7 +13,7 @@
   </developers>
   <inceptionYear>2017</inceptionYear>
   <properties>
-    <java.version>1.8</java.version>
+    <java.version>11</java.version>
     <maven-compiler-plugin.version>3.5.1</maven-compiler-plugin.version>
   </properties>
   <build>
