+++ src/main/java/module-info.java
@@ -1,0 +1,8 @@
+module kn.uni.dbis.cs.sorting {
+    exports kn.uni.dbis.cs.sorting;
+    exports kn.uni.dbis.cs.sorting.algo;
+    exports kn.uni.dbis.cs.sorting.gui;
+
+    // for awt and swing modules
+    requires java.desktop;
+}
